# [PROF-11898] Switch profiler stack truncation strategy and improve sampling performance