--- a/benchmarks/profiling_sample_loop_v2.rb
+++ b/benchmarks/profiling_sample_loop_v2.rb
@@ -43,16 +43,47 @@ class ProfilerSampleLoopBenchmark
     Thread.new { deep_stack.call }.tap { |t| t.name = "Deep stack #{depth}" }
   end
 
+  def go_to_depth_and_run(depth: 2900, &block)
+    current_depth = caller.size
+    return yield if current_depth > depth
+
+    # rubocop:disable Lint/DuplicateBranch
+    # Simulate "some" complexity in the method bytecode
+    case current_depth % 10
+    when 0
+      go_to_depth_and_run(depth: depth, &block)
+    when 1
+      go_to_depth_and_run(depth: depth, &block)
+    when 2
+      go_to_depth_and_run(depth: depth, &block)
+    when 3
+      go_to_depth_and_run(depth: depth, &block)
+    when 4
+      go_to_depth_and_run(depth: depth, &block)
+    when 5
+      go_to_depth_and_run(depth: depth, &block)
+    when 6
+      go_to_depth_and_run(depth: depth, &block)
+    when 7
+      go_to_depth_and_run(depth: depth, &block)
+    when 8
+      go_to_depth_and_run(depth: depth, &block)
+    when 9
+      go_to_depth_and_run(depth: depth, &block)
+    end
+    # rubocop:enable Lint/DuplicateBranch
+  end
+
   def run_benchmark(mode: :ruby)
     threads = Array.new(4) { mode == :ruby ? thread_with_very_deep_stack : thread_with_very_deep_stack_and_native_frames }
     collector = Datadog::Profiling::Collectors::ThreadContext.for_testing(recorder: @recorder)
 
     if mode == :native
-      if !Datadog::Profiling::Collectors::Stack._native_filenames_available?
+      unless Datadog::Profiling::Collectors::Stack._native_filenames_available?
         if OS.linux?
           raise 'Native filenames are not available. This is not expected on Linux!'
         else
-          puts "Skipping benchmarking native_frames, not supported outside of Linux"
+          puts 'Skipping benchmarking native_frames, not supported outside of Linux'
           return
         end
       end
@@ -70,21 +101,11 @@ class ProfilerSampleLoopBenchmark
         **benchmark_time,
       )
 
-      x.report("stack collector (#{mode} frames - native filenames enabled) #{ENV['CONFIG']}") do
-        Datadog::Profiling::Collectors::ThreadContext::Testing._native_sample(
-          collector,
-          PROFILER_OVERHEAD_STACK_THREAD,
-          false
-        )
-      end
+      x.report("stack collector (#{mode} frames - native filenames enabled) #{ENV['CONFIG']}") { sample(collector) }
 
       if mode == :native
         x.report("stack collector (#{mode} frames - native filenames disabled) #{ENV['CONFIG']}") do
-          Datadog::Profiling::Collectors::ThreadContext::Testing._native_sample(
-            collector_without_native_filenames,
-            PROFILER_OVERHEAD_STACK_THREAD,
-            false
-          )
+          sample(collector_without_native_filenames)
         end
       end
 
@@ -95,6 +116,40 @@ class ProfilerSampleLoopBenchmark
     threads.map(&:kill).each(&:join)
     @recorder.serialize!
   end
+
+  def run_varying_depth_benchmark
+    collector = Datadog::Profiling::Collectors::ThreadContext.for_testing(recorder: @recorder, max_frames: 3000)
+
+    Benchmark.ips do |x|
+      benchmark_time = VALIDATE_BENCHMARK_MODE ? { time: 0.01, warmup: 0 } : { time: 10, warmup: 2 }
+      x.config(
+        **benchmark_time,
+      )
+
+      # This benchmark checks the performance of samples when the stack keeps changing
+      x.report("stack collector (varying depth) #{ENV['CONFIG']}") do
+        sample(collector)
+        add_extra_frame_and_sample(collector) # This makes the stack change
+      end
+
+      x.save! "#{File.basename(__FILE__)}-results.json" unless VALIDATE_BENCHMARK_MODE
+      x.compare!
+    end
+
+    File.write('test.out', @recorder.serialize!._native_bytes)
+  end
+
+  def sample(collector)
+    Datadog::Profiling::Collectors::ThreadContext::Testing._native_sample(
+      collector,
+      PROFILER_OVERHEAD_STACK_THREAD,
+      false
+    )
+  end
+
+  def add_extra_frame_and_sample(collector)
+    sample(collector)
+  end
 end
 
 puts "Current pid is #{Process.pid}"
@@ -103,4 +158,5 @@ ProfilerSampleLoopBenchmark.new.instance_exec do
   create_profiler
   run_benchmark(mode: :ruby)
   run_benchmark(mode: :native)
+  go_to_depth_and_run { run_varying_depth_benchmark }
 end