--- a/benchmarks/profiling_sample_loop_v2.rb
+++ b/benchmarks/profiling_sample_loop_v2.rb
@@ -136,7 +136,7 @@ class ProfilerSampleLoopBenchmark
       x.compare!
     end
 
-    File.write('test.out', @recorder.serialize!._native_bytes)
+    @recorder.serialize!
   end
 
   def sample(collector)